backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
yfinance==0.2.18
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
textblob==0.17.1
